fix: change type to float64 for healthvalue, add more coverage testing, fix return values for missing healthValue

+++ backend/internal/db/repo_health_threshold.go
@@ -9,8 +9,8 @@ import (
 )
 
 type Threshold struct {
-	Upper float32 `json:"upper"`
-	Lower float32 `json:"lower"`
+	Upper float64 `json:"upper"`
+	Lower float64 `json:"lower"`
 }
 
 type RepoHealthThreshold struct {
